requirements.txt
python-escpos
requests
win32printing
arrow
pyinstaller
